requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.27.0
pyyaml
